Port engine to arm64 macOS and SDL2
Adds an arm64 target, 64-bit safe VM/syscall signatures, and migrates the
four SDL sources from the end-of-life SDL 1.2 to SDL2.

Links real SDL2 by versioned Cellar path: /opt/homebrew/opt/sdl2 is itself a
symlink to sdl2-compat, and the real keg's install name still points there, so
the link is verified after every build.

+++ Game/Game/g_main.c
@@ -216,33 +216,33 @@ This is the only way control passes into the module.
 This must be the very first function compiled into the .q3vm file
 ================
 */
-Q_EXPORT intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
-	switch ( command ) {
+Q_EXPORT intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6, intptr_t arg7, intptr_t arg8, intptr_t arg9, intptr_t arg10, intptr_t arg11  ) {
+	switch ( (int)command ) {
 		case GAME_INIT:
-			G_InitGame( arg0, arg1, arg2 );
+			G_InitGame( (int)arg0, (int)arg1, (int)arg2 );
 			return 0;
 		case GAME_SHUTDOWN:
-			G_ShutdownGame( arg0 );
+			G_ShutdownGame( (int)arg0 );
 			return 0;
 		case GAME_CLIENT_CONNECT:
-		return (intptr_t)ClientConnect( arg0, arg1 );
+			return (intptr_t)ClientConnect( (int)arg0, (int)arg1 );
 		case GAME_CLIENT_THINK:
-			ClientThink( arg0 );
+			ClientThink( (int)arg0 );
 			return 0;
 		case GAME_CLIENT_USERINFO_CHANGED:
-			ClientUserinfoChanged( arg0 );
+			ClientUserinfoChanged( (int)arg0 );
 			return 0;
 		case GAME_CLIENT_DISCONNECT:
-			ClientDisconnect( arg0 );
+			ClientDisconnect( (int)arg0 );
 			return 0;
 		case GAME_CLIENT_BEGIN:
-			ClientBegin( arg0 );
+			ClientBegin( (int)arg0 );
 			return 0;
 		case GAME_CLIENT_COMMAND:
-			ClientCommand( arg0 );
+			ClientCommand( (int)arg0 );
 			return 0;
 		case GAME_RUN_FRAME:
-			G_RunFrame( arg0 );
+			G_RunFrame( (int)arg0 );
 			return 0;
 		case GAME_CONSOLE_COMMAND:
 			return ConsoleCommand();
